\name{BonusLab-package}
\alias{BonusLab-package}
\alias{BonusLab}
\docType{package}
\title{
\packageTitle{BonusLab}
}
\description{
\packageDescription{BonusLab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BonusLab}
\packageIndices{BonusLab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BonusLab}

Maintainer: \packageMaintainer{BonusLab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
