% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BonusLab-package.R
\docType{package}
\name{BonusLab}
\alias{BonusLab}
\title{Bonus package}
\description{
Ridge regression, airport delay visualizations and predictions.
}
\author{
Leonard Persson Norblad \email{leope744@student.liu.se} Jacob Sibanda \email{jacsi259@student.liu.se}
}
