\name{lab4package}
\alias{lab4package}
\docType{package}
\title{Lab 4 solution from ductr388 and wilwi856}
\description{
This packages fits a multiple linear regression model with ordinary least squares. 
The package creates an reference class object that stores the fitted model.
The following methods are available: print(), pred(), coef(), summary(), and plot(). 
}
\details{
Package: lab4package

Type: Package

Version: 1.0.0

Date: 2023-09-25

Licence: GPL (>=3)


}
\author{

Duc Tran Maintainer: <ductr388@student.liu.se> 

William Wiik Creator: <wilwi856@student.liu.se>
}
\references{
Information about the ordinary least squares can be found 
\href{https://en.wikipedia.org/wiki/Ordinary_least_squares}{here}.
}



