% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{Refers to the object return by the linear regression function}

\item{...}{arguments to be passed to methods}
}
\value{
object$coefficients
}
\description{
Returns the coefficients
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          coef(s)
}
