% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BonusLab.R
\docType{package}
\name{lab6}
\alias{lab6}
\title{Bonus Lab}
\description{
Using Ridge regression to improve lab 4
}
\author{
Kyriakos Papadopoulos  \email{kyrpa853@student.liu.se} YiHung Chen  \email{yihch883@student.liu.se}
}
