% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ref_class.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg_ref_class.R}
\description{
Reference class for the function linreg
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula used to fit the regression.}

\item{\code{input_arg_data}}{The input used for the data argument.}

\item{\code{coefficients}}{Coefficients for the regression model.}

\item{\code{fitted_values}}{Fitted values for the model.}

\item{\code{residuals}}{Residuals for the model.}

\item{\code{df}}{Degrees of freedoms for the model.}

\item{\code{residual_variance}}{The variance of the residuals.}

\item{\code{coefficients_variance}}{Covariance matrix of the regression coefficients.}

\item{\code{t_values}}{t-values for the coefficients.}

\item{\code{p_values}}{p-values for the coefficients.}
}}


