% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear regression functions}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{Y dependent variable and X independent variable of the data frame}

\item{data}{Data frame input by the user}
}
\value{
res
}
\description{
Linear regression functions
}
\examples{
data(iris)
          linreg(Petal.Length~Species, iris)
}
