% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ref_class.R
\name{linreg_coef}
\alias{linreg_coef}
\title{Coefficients for linreg object.}
\value{
A print of the coefficents from the regression model.
}
\description{
The estimated coefficients from the regression model for linreg objects.
}
