% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ref_class.R
\name{linreg_pred}
\alias{linreg_pred}
\title{Predicted values for linreg object.}
\value{
A named vector with the fitted values from the regression model.
}
\description{
Fitted values from the regression model of reference class linreg.
}
