% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ref_class.R
\name{linreg_print}
\alias{linreg_print}
\title{Print function for linreg object.}
\value{
A printout of the function call and coefficients from the regression model.
}
\description{
Print method for the reference class linreg.
}
