% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_qr.R
\name{linreg_qr}
\alias{linreg_qr}
\title{Linreg_qr}
\usage{
linreg_qr(formula, data)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a dataframe containing the variables in the formula}
}
\value{
returns object of class "linreg"
}
\description{
Linreg_qr is used to carry out multiple linear regression, with qr factorization
to estimate the coeffcients and the variance of the coefficients.
}
\examples{
data(iris)
linres <- linreg(formula=Sepal.Length ~ Petal.Length + Petal.Width, data=iris)
}
