% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_qr_decomp.R
\name{linreg_qr_decomp}
\alias{linreg_qr_decomp}
\title{Linear regression using QR decomposition}
\usage{
linreg_qr_decomp(data, formula)
}
\arguments{
\item{data}{A data frame containg all the data}

\item{formula}{The regression formula}
}
\value{
The estimated coefficients and their variance
}
\description{
Linear regression using QR decomposition
}
\examples{
linreg_qr_decomp(iris, as.formula("Petal.Length ~ Species"))
linreg_qr_decomp(iris, as.formula("Petal.Length ~ Sepal.Width + Sepal.Length"))
}
\references{
\url{https://en.wikipedia.org/wiki/QR_decomposition}
}
