% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ref_class.R
\name{linreg_resid}
\alias{linreg_resid}
\title{Residuals for linreg object.}
\value{
A named vector with the residuals from the regression model.
}
\description{
Residuals from the regression model of reference class linreg.
}
