% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ref_class.R
\name{linreg_summary}
\alias{linreg_summary}
\title{Summary for linreg object.}
\value{
A print of the coefficents from the regression model with their 
respective standard deviation, t-values, and p-values.
The models residual standard error and degrees of freedom are also printed.

Two different ggplot objects.
}
\description{
A short summary of a linreg object.

Produces a residual vs fitted plot and a scale-location plot.
The scale-location plot shows square roof of standardized residuals vs fitted values.
}
