% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{make_plot}
\alias{make_plot}
\title{Make_plot}
\usage{
make_plot(
  df,
  title,
  cpt,
  xaxis,
  yaxis,
  xlab,
  ylab,
  cutoff_y_var,
  cutoff_y_val,
  id
)
}
\arguments{
\item{df}{data frame that contains the data to be plotted}

\item{title}{title of the plot}

\item{cpt}{caption}

\item{xaxis}{x axis variable}

\item{yaxis}{y axis variable}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{cutoff_y_var}{cutoff variable for which the cutoff_y_val should be
applied (see cutoff_y_val)}

\item{cutoff_y_val}{cutoff value for which id labels should be displayed
for the variable as given in cutoff_y_var}

\item{id}{variable that holds the id values of the observations}
}
\value{
ggplot object
}
\description{
Help function to the plot.linreg function
}
