% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BonusLab.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{mtcars}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\usage{
mtcars
}
\description{
A Graph(data.frame) contain boston housing info
}
\keyword{datasets}
