% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{pred}
\alias{pred}
\title{Predicted values}
\usage{
pred(object)
}
\arguments{
\item{object}{Refers to the object return by the linear regression function}
}
\value{
object$fitted_values
}
\description{
Returns the predicted values
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          pred(s)
}
