% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rr.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Residual of the linear regression function}
\usage{
\method{predict}{ridgereg}(object, x, ...)
}
\arguments{
\item{object}{Refers to the object return by the linear regression function}

\item{...}{arguments to be passed to methods}
}
\value{
object$residuals

predicted values
}
\description{
Returns the residuals
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          resid(s)
Predicted values

Returns the predicted values

data(iris)
          res<-ridgereg(formula=Petal.Length~Species, data= iris)
          pred(res)
}
