% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rr.R
\name{predr}
\alias{predr}
\title{Predicted values}
\usage{
predr(object, x, ...)
}
\arguments{
\item{object}{Refers to the object return by the linear regression function}

\item{x}{new data to be predicted}

\item{...}{arguments to be passed to methods}
}
\value{
predicted values
}
\description{
Returns the predicted values
}
\examples{
data(iris)
          res<-ridgereg(formula=Petal.Length~Species, data= iris, lambda=2)
          pred(res)
}
