% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{print.linreg}
\alias{print.linreg}
\title{Print}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{Refers to the object return by the linear regression function}

\item{...}{arguments to be passed to methods}
}
\value{
Call, Coefficients
}
\description{
Prints model call and coefficients
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data=iris)
          print(s)

}
