% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rr.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{Refers to the object return by the linear regression function}

\item{...}{arguments to be passed to methods}
}
\value{
Call, Coefficients
}
\description{
Prints model call and coefficients
}
\examples{
data(iris)
          s<-ridgereg(formula=Petal.Length~Species, data=iris,lambda=2)
          print(s)

}
