% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_hr.R
\name{qr_hr}
\alias{qr_hr}
\title{QR factorization with Householder reflection}
\usage{
qr_hr(A)
}
\arguments{
\item{A}{n x p (p <= n) matrix for which to do the QR factorization}
}
\value{
list with Q and R matrices
}
\description{
This function performs QR factorisation with Householder reflection.
See e.g. Some	Notes	on	Least	Squares, QR-factorization, SVD and
Fitting. Ove	Edlund, Department of Science and Mathematics
January 23, 2013. C0002M Numerical Analysis.
}
\examples{
data(iris)
formula <- Sepal.Length ~ Petal.Length + Petal.Width
data <- iris
X <-stats::model.matrix(object=formula, data=data)
QR <- qr_hr(X)
}
