% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRridgereg.R
\docType{class}
\name{QRridgereg-class}
\alias{QRridgereg-class}
\alias{QRridgereg}
\title{A reference class to perform ridge regression with QR decomposition.}
\description{
A reference class to perform ridge regression with QR decomposition.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula: Object dependent and indepedent variables.}

\item{\code{data}}{data.frame: All the data.}

\item{\code{dname}}{vector: The names of the variables.}

\item{\code{lambda:}}{parameter lambda}

\item{\code{X}}{matrix: Independent variables.}

\item{\code{y}}{matrix: Dependent variable.}

\item{\code{b_ridge}}{matrix: Estimation of the regression coefficients.}

\item{\code{y_hat}}{matrix: Estimation of the y values.}

\item{\code{e_hat}}{matrix: Estimation of the error variable i.e. the residuals.}

\item{\code{m}}{vector: mean of each covariate.}

\item{\code{s}}{vector: sd of each covariate.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Return the coefficient}

\item{\code{initialize(formula, data, lambda)}}{Initialize the values}

\item{\code{predict(newdata = NA)}}{Return the predicted values}

\item{\code{print()}}{Print the coeff that needed}
}}

