% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residual of the linear regression function}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{Refers to the res variable return by the linear regression function}

\item{...}{arguments to be passed to methods}
}
\value{
object$residuals
}
\description{
Returns the residuals
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          resid(s)
}
