% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridgereg.R
\docType{class}
\name{Ridgereg-class}
\alias{Ridgereg-class}
\alias{Ridgereg}
\title{BonusLab: Computations using ordinary least squares, implement methods in RC class}
\arguments{
\item{FunctionFormula}{formula object passed by user}

\item{data}{dataframe passed by the user}

\item{lambda}{value}

\item{l}{value of Lambda}

\item{y}{dependent variables matrix}

\item{RegCofMatrix}{independent variables matrix}

\item{X}{independent variables matrix}
}
\value{
class

the value that has been set
Setter for Fitted Values

the value that has been set
Setter for Fitted Values

the value that has been set
Setter for degree of freedom

the value that has been set
Setter for degree of residual variance

the value that has been set
Setter for degree of variance of reg coefficients

the value that has been set
Setter T Values

the value that has been set
Setter P Values

the value that has been set
Print like lm

nothing
Plot using ggplot2

nothing
Residuals calculation

Residuals
predicted values

FittedValues
coefficeints as named vec

coefficeints as named vec
Print the Summary

coefficeints as named vec
}
\description{
Sets the value for RegressionCoeficientMatrix field

Sets the value for FittedValues

Sets the value for Residuals

Sets the value for DegreesOfFreedom

Sets the value for ResidualVariance

Sets the value for VarianceOfTheRegressionCoefficients

Sets the value for TValues

Sets the value for TValues

Prints out the coefficient and coefficent names

Plots 2 graphs on a grid that are mention in the lab manual

return only residuals as vec form

return the preticted values (y-hat)

return the coefficeints as named vec

Print the simmary along with the pvalue, tvalue, sigma and degree of freedom
}
\section{Fields}{

\describe{
\item{\code{ridge_VarianceOfTheRegressionCoefficients}}{matrix.}

\item{\code{ridge_RegressionCoeficientMatrix}}{matrix.}

\item{\code{ridge_TValues}}{t-values in vec form}

\item{\code{ridge_Residuals}}{matrix.}

\item{\code{ridge_ResidualVariance}}{matrix.}

\item{\code{ridge_DegreesOfFreedom}}{numeric.}

\item{\code{ridge_DataName}}{Name of the dataframe passed.}

\item{\code{ridge_FittedValues}}{matrix.}

\item{\code{ridge_PValues}}{t-values in vec form}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{return the coefficeints as named vec}

\item{\code{initialize(formula, data, lam)}}{Constructor to initialize the data fields of this class}

\item{\code{plot()}}{Plots 2 graphs on a grid that are mention in the lab manual}

\item{\code{pred()}}{return the preticted values (y-hat)}

\item{\code{print()}}{Prints out the coefficient and coefficent names}

\item{\code{resid()}}{return only residuals as vec form}

\item{\code{ridge_setDegreesOfFreedom(RegCofMatrix, X)}}{Sets the value for DegreesOfFreedom}

\item{\code{ridge_setFittedValues(RegCofMatrix, X)}}{Sets the value for FittedValues}

\item{\code{ridge_setPValues()}}{Sets the value for TValues}

\item{\code{ridge_setRegCofficient(X, y, l)}}{Sets the value for RegressionCoeficientMatrix field}

\item{\code{ridge_setResidual(y)}}{Sets the value for Residuals}

\item{\code{ridge_setResidualVariance()}}{Sets the value for ResidualVariance}

\item{\code{ridge_setTValues()}}{Sets the value for TValues}

\item{\code{ridge_setVarianceOfTheRegressionCoefficients(X)}}{Sets the value for VarianceOfTheRegressionCoefficients}

\item{\code{summary()}}{Print the simmary along with the pvalue, tvalue, sigma and degree of freedom}
}}

\examples{
newobj= Ridgereg$new(frm, iris,3)
newobj$summary()

}
