% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg-package.R
\docType{package}
\name{ridgereg-package}
\alias{ridgereg}
\alias{ridgereg-package}
\title{ridgereg: This package is for R Programming Advanced ML bonus lab which is a multiple regression model}
\description{
This is a Multiple Linear Regression model.
}
\author{
\strong{Maintainer}: Xuan Developer \email{xuawa284@student.liu.se}

Other contributors:
\itemize{
  \item Ashen Developer \email{thawe276@student.liu.se} [contributor]
}

}
\keyword{internal}
