% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{ridgereg}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a dataframe containing the variables in the formula}

\item{lambda}{regularisation parameter}
}
\value{
returns object of class "ridgereg"
}
\description{
ridgereg is used to carry out ridge regression.
}
\examples{
data(iris)
ridgeres <- ridgereg(formula=Sepal.Length ~ Petal.Length + Petal.Width,
data=iris, lambda=2)
}
