% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_ref_class.R
\name{ridgereg_coef}
\alias{ridgereg_coef}
\title{Coefficients for ridgereg object.}
\value{
A print of the coefficents from the ridge regression model.
}
\description{
The estimated coefficients from the regression model for ridgereg objects.
}
