% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_ref_class.R
\name{ridgereg_predict}
\alias{ridgereg_predict}
\title{Predict function for ridgereg object.}
\value{
A vector of the predicted values.
}
\description{
Predict new values of observation with the existing ridge regression model.
}
