% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_ref_class.R
\name{ridgereg_print}
\alias{ridgereg_print}
\title{Print function for ridgereg object.}
\value{
A printout of the function call and coefficients from the regression model.

A printout of the function call and coefficients from the regression model.
}
\description{
Print method for the reference class ridgereg.

Show method for the reference class ridgereg.
}
