% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_Decomposition.R
\name{ridgereg_QR}
\alias{ridgereg_QR}
\title{BonusLab: Computations using QR Decomposition, implement method}
\usage{
ridgereg_QR(formula, data, lambda)
}
\arguments{
\item{data}{dataframe passed by the user}

\item{lambda}{value}

\item{Formula}{formula object passed by user}
}
\description{
BonusLab: Computations using QR Decomposition, implement method
}
\examples{
newobj= Ridgereg(frm, iris,3)
}
