% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rr.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(error)
}
\arguments{
\item{error}{residuals in model}
}
\value{
rmse value
}
\description{
RMSE
}
\examples{
data(iris)
ridgeres <- ridgereg(formula=Sepal.Length ~ Petal.Length + Petal.Width,
data=iris, lambda=2)
x = data.frame(Petal.Length=iris$Petal.Length,Petal.Width=iris$Petal.Width)
rr_error <- iris$Sepal.Length-predict(ridgeres,x)
rmse_ridge <- rmse(rr_error)
}
