% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary of linear regression function}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{Refers to the object return by the linear regression function}

\item{...}{arguments to be passed to methods}
}
\value{
data frame with the final computations
}
\description{
Returns a summary of the results from the linear regression: estimated
coefficients, their standard errors, tvalues and p-values, as well as
the residual variance.
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          summary(s)
}
