% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Mean Airport Delays}
\usage{
visualize_airport_delays()
}
\value{
A ggplot object.
}
\description{
Creates a plot visualizing the mean arrival delay of flights for different
NYC airports, plotted by their longitude and latitude.
This function uses data from the \code{nycflights13} package.
}
\examples{
\dontrun{
  visualize_airport_delays()
}
}
