% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_ref_class.R
\docType{package}
\name{bonuslabpackage-package}
\alias{bonuslabpackage}
\alias{bonuslabpackage-package}
\title{bonuslabpackage: Linear regression with OLS and ridge.}
\description{
Contains two RC-classes, "linreg" that fits a linear regression and "ridgereg" that fits a ridge regression.
}
\author{
\strong{Maintainer}: Duc Tran \email{ductr388@student.liu.se}

Authors:
\itemize{
  \item William Wiik \email{wilwi856@student.liu.se}
}

}
\keyword{internal}
