% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg is to build a multiple regression mode.}
\arguments{
\item{formula}{a formula specifying the regression model, e.g., y ~ x1 + x2}

\item{data}{a data frame containing the variables specified in the formula.}
}
\value{
An object of class "linreg" containing regression results.
}
\description{
linreg is to build a multiple regression mode based on a given dataset and a formula.
}

\examples{
data(iris)
linreg_mod_object <- linreg$new(Petal.Length~Species, data = iris)
linreg_mod_object$print()
linreg_mod_object$plot()
linreg_mod_object$resid()
linreg_mod_object$pred()
linreg_mod_object$coef()
linreg_mod_object$summary()
}
\references{
Reference page link <https://dataxujing.github.io/R_oop/RC.html#section-5.5.1>
}
