% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A reference class representing av linear regression model}
\value{
An object of class linreg
}
\description{
A reference class representing av linear regression model
}
\section{Fields}{

\describe{
\item{\code{beta_hat}}{A matrix containing the estimated coefficients}

\item{\code{y_hat}}{A matrix containing the fitted values}

\item{\code{res}}{A matrix containing the residuals}

\item{\code{df}}{A numeric vector containg the degrees of freedom}

\item{\code{res_var}}{A numeric vector containing the residual variance}

\item{\code{beta_hat_var}}{A matrix containing the variance of the estimated coefficients}

\item{\code{t_values}}{A matrix containg the t-values of the estimated coefficients}

\item{\code{p_values}}{A matrix containg the p-values of the estimated coefficients}

\item{\code{resstd}}{A matrix containg the standardized residuals}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the estimated coefficients}

\item{\code{plot()}}{Plots the residuals versus the fitted values and the square root of the standardized residuals versus the fitted values}

\item{\code{pred()}}{Returns a vector with the fitted values}

\item{\code{print()}}{Prints out the coefficients and the coefficient names}

\item{\code{resid()}}{Returns a vector with the residuals}

\item{\code{summary()}}{Returns a summary of the linear regression model}
}}

\examples{
linreg(iris, as.formula("Petal.Length ~ Species"))
linreg(iris, as.formula("Petal.Length ~ Sepal.Width + Sepal.Length"))
}
\references{
\url{https://en.wikipedia.org/wiki/Least_squares}
}
