% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression - Function for a multiple regression model}
\arguments{
\item{formula}{an object of class 'formula': a symbolic description of the model to be fitted.}

\item{data}{an 'data.frame' containing the variables in the model.}

\item{lambda}{a penalty parameter (numeric), normally small (sowhere between 0 and 0.1)}
}
\value{
Returns an object of class 'ridgereg'.
}
\description{
This functions estimates the coefficients and fitted values by optimizing a penalized least square problem.
}

\examples{
data("iris")
library(BonusLab)
ridgereg <- getFromNamespace("ridgereg", "BonusLab")
ridgereg_model <- ridgereg(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda=0.01)
}
