% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{ridgereg is to build a ridge regression mode.}
\arguments{
\item{formula}{a formula specifying the regression model, e.g., y ~ x1 + x2}

\item{data}{a data frame containing the variables specified in the formula.}

\item{lambda}{the variable λ specifying the variable for ridge  regression formula.}
}
\value{
An object of class "ridgereg" containing ridge regression results.
}
\description{
ridgereg is to build a multiple regression mode based on a given dataset and a formula.
}

\examples{
data(iris)
ridgereg_mod_object <- ridgereg$new(Petal.Length~Species, data = iris, lambda=0)
ridgereg_mod_object$print()
ridgereg_mod_object$resid()
ridgereg_mod_object$pred()
ridgereg_mod_object$coef()
ridgereg_mod_object$summary()
}
\references{
Reference page link <https://dataxujing.github.io/R_oop/RC.html#section-5.5.1>
}
