% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge regression}
\arguments{
\item{Data}{A data.frame}

\item{formula}{A formula}

\item{lambda}{Regularization parameter}

\item{new_data}{New data to be predicted in the \code{predict()} function}
}
\value{
\code{print()} Returns the function call and estimated coefficients

\code{predict()} Returns predictions for new data

\code{coef()} Returns a vector of estimated coefficients
}
\description{
Performs ridge regression
}

\examples{
data(iris)
example <- ridgereg(data=iris, formula=Petal.Length~Petal.Width+Sepal.Width+Species, lambda=0)
example$print()
example$predict(new_data=data.frame(Petal.Width=c(1,2,3),
                                   Sepal.Width=c(2,3,4),
                                   Species=c("versicolor", "virginica", "versicolor")))
}
