% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge regression with QR}
\description{
Ridge regression with QR
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{lambda}}{numeric.}

\item{\code{parsedata}}{character.}

\item{\code{names}}{vector.}

\item{\code{beta}}{numeric.}

\item{\code{data_n}}{character.}
}}


