% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{A reference class representing av ridge regression model}
\value{
An object of class linreg
}
\description{
A reference class representing av ridge regression model
}
\section{Fields}{

\describe{
\item{\code{beta_hat}}{A matrix containing the estimated coefficients}

\item{\code{y_hat}}{A matrix containing the fitted values}

\item{\code{res}}{A matrix containing the residuals}

\item{\code{df}}{A numeric vector containg the degrees of freedom}

\item{\code{res_var}}{A numeric vector containing the residual variance}

\item{\code{beta_hat_var}}{A matrix containing the variance of the estimated coefficients}

\item{\code{t_values}}{A matrix containg the t-values of the estimated coefficients}

\item{\code{p_values}}{A matrix containg the p-values of the estimated coefficients}

\item{\code{resstd}}{A matrix containg the standardized residuals}

\item{\code{reg_formula}}{A formula containing the regression formula}

\item{\code{data_name}}{A character vector containing the name of the data frame}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the estimated coefficients}

\item{\code{plot()}}{Plots the residuals versus the fitted values and the square root of the standardized residuals versus the fitted values}

\item{\code{predict(object, newdata)}}{Returns the predicted values}

\item{\code{print()}}{Prints out the coefficients and the coefficient names}

\item{\code{resid()}}{Returns a vector with the residuals}

\item{\code{summary()}}{Returns a summary of the linear regression model}
}}

\examples{
ridgereg(iris, as.formula("Petal.Length ~ Species"))
ridgereg(iris, as.formula("Petal.Length ~ Sepal.Width + Sepal.Length"))
}
\references{
\url{https://en.wikipedia.org/wiki/Ridge_regression}
}
