% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_ref_class.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Reference class for the function ridgereg}
\description{
Reference class for the function ridgereg
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula used to fit the regression.}

\item{\code{input_arg_data}}{The input used for the data argument.}

\item{\code{coefficients}}{Coefficients for the regression model.}

\item{\code{fitted_values}}{Fitted values for the model.}

\item{\code{lambda}}{Penalty parameter for the ridge regression.}

\item{\code{coef_names}}{Names of the coefficients from the ridge regression.}

\item{\code{mean_X}}{Mean value for each of the dependent variables.}

\item{\code{sd_X}}{Standard deviation for each of the dependent variables.}

\item{\code{mean_y}}{Mean value for the dependent variable.}
}}


