% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression}
\value{
Returns an object of the class 'ridgereg'
}
\description{
Sets reference class for ridge regression model
}
\section{Fields}{

\describe{
\item{\code{formula}}{as a formula}

\item{\code{data}}{as a data frame}

\item{\code{norm_matrix}}{as a matrix}

\item{\code{lambda}}{as numeric}

\item{\code{y_hat}}{as a matrix}

\item{\code{beta_hat}}{as a vector}
}}


