% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rifge.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge regression}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{The Y and X variables of the regression}

\item{data}{A data frame}

\item{lambda}{Hyperparameter of the regression}
}
\description{
Ridge regression
}
