% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Airport delays}
\usage{
visualize_airport_delays()
}
\value{
Returns a plot
}
\description{
This function creates a plot that visualizes the mean delay of  fights for different airports by longitude and latitude
}
\examples{
library(BonusLab)
plot <- BonusLab::visualize_airport_delays()
}
