% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Airport Delays}
\usage{
visualize_airport_delays()
}
\value{
A ggplot2 plot visualizing the mean flight delay by airport location.
}
\description{
This function creates a plot that visualizes the mean delay of flights for different airports by longitude and latitude using ggplot2.
}
\details{
The function loads the nycflights13 dataset, filters out missing latitude and longitude values, calculates the mean delay for each airport, and then creates a scatterplot.
}
\examples{
visualize_airport_delays()
}
\seealso{
\code{\link{ggplot2}}

\code{\link{nycflights13}}
}
