% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Airport delays}
\usage{
visualize_airport_delays()
}
\value{
A visual representation of a map in a form of a scatterplot of airport delays.
}
\description{
Visualize airport departure delays for US airports
}
\examples{
visualize_airport_delays()

}
