% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualizes the mean delay of flights for different airports}
\usage{
visualize_airport_delays()
}
\value{
link map and heat map to visual airplanes traffic delays
}
\description{
Visualizes the mean delay of flights for different airports
}
\examples{
visualize_airport_delays()
}
