% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualisation of airport delays}
\usage{
visualize_airport_delays()
}
\value{
A ggplot visualising the mean delays.
}
\description{
Visualizes the mean delay of flights for different airports by longitude and latitude
from the dataset in the package nycflights13.
}
