% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Airport Delays}
\usage{
visualize_airport_delays()
}
\value{
Returns a graph
}
\description{
This function creates a plot that visualizes the mean delay of flights for different airports by longitude and latitude.
}
